Try to get network connection on start-up
If the device is not online when starting the application, open the network
connectivity dialog. The opening is delayed until the QML app is shown to not
confuse the user by showing the network dialog for a short time, then showing
the app, thus hiding the dialog.

Also the dialog is shown only once since GUPnP can handle appearing and
disappearing network interfaces quite fine on its own.